+++ backend/src/main/java/generator/config/BeanConfig.java
@@ -1,0 +1,61 @@
+package generator.config;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.module.SimpleModule;
+import com.fasterxml.jackson.databind.ser.std.DateSerializer;
+import com.fasterxml.jackson.databind.ser.std.ToStringSerializer;
+import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.support.ReloadableResourceBundleMessageSource;
+
+import java.text.SimpleDateFormat;
+import java.util.Date;
+import java.util.TimeZone;
+
+import generator.constant.CommonConstant;
+
+/**
+ * BeanConfig.
+ *
+ * @author 刘斌
+ * @version 0.0.1
+ * @serial 2022-08-16 : base version.
+ */
+@Configuration
+public class BeanConfig {
+    /**
+     * jackSon 的 ObjectMapper
+     *
+     * @return ObjectMapper
+     */
+    @Bean
+    public ObjectMapper objectMapper() {
+        // 序列化设置
+        final ObjectMapper objectMapper = new ObjectMapper();
+        objectMapper.setTimeZone(TimeZone.getDefault());
+        // 序列换成json时,将所有的long变成string
+        SimpleModule simpleModule = new SimpleModule();
+        simpleModule.addSerializer(Long.class, ToStringSerializer.instance);
+        simpleModule.addSerializer(Long.TYPE, ToStringSerializer.instance);
+        // 日期序列化设置
+        JavaTimeModule javaTimeModule = new JavaTimeModule();
+        javaTimeModule.addSerializer(Date.class, new DateSerializer(false, new SimpleDateFormat(CommonConstant.DATE_TIME_FORMATTER_STRING)));
+        objectMapper.registerModule(simpleModule).registerModule(javaTimeModule);
+
+        return objectMapper;
+    }
+
+    /**
+     * 国际化
+     *
+     * @return 对象
+     */
+    @Bean
+    public ReloadableResourceBundleMessageSource messageSource() {
+        ReloadableResourceBundleMessageSource messageSource = new ReloadableResourceBundleMessageSource();
+        messageSource.setBasename("classpath:org/springframework/security/messages_zh_CN");
+        return messageSource;
+    }
+}
