+++ backend/src/main/java/generator/config/WebMvcConfig.java
@@ -1,20 +1,10 @@
 package generator.config;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.module.SimpleModule;
-import com.fasterxml.jackson.databind.ser.std.DateSerializer;
-import com.fasterxml.jackson.databind.ser.std.ToStringSerializer;
-import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.converter.json.MappingJackson2HttpMessageConverter;
-
-import java.text.SimpleDateFormat;
-import java.util.Date;
-import java.util.TimeZone;
-
-import generator.constant.CommonConstant;
 
 /**
  * WebMvcConfig.
@@ -25,28 +15,6 @@ import generator.constant.CommonConstant;
  */
 @Configuration
 public class WebMvcConfig {
-
-    /**
-     * jackSon 的 ObjectMapper
-     *
-     * @return ObjectMapper
-     */
-    @Bean
-    public ObjectMapper objectMapper() {
-        // 序列化设置
-        final ObjectMapper objectMapper = new ObjectMapper();
-        objectMapper.setTimeZone(TimeZone.getDefault());
-        // 序列换成json时,将所有的long变成string
-        SimpleModule simpleModule = new SimpleModule();
-        simpleModule.addSerializer(Long.class, ToStringSerializer.instance);
-        simpleModule.addSerializer(Long.TYPE, ToStringSerializer.instance);
-        // 日期序列化设置
-        JavaTimeModule javaTimeModule = new JavaTimeModule();
-        javaTimeModule.addSerializer(Date.class, new DateSerializer(false, new SimpleDateFormat(CommonConstant.DATE_TIME_FORMATTER_STRING)));
-        objectMapper.registerModule(simpleModule)
-                .registerModule(javaTimeModule);
-        return objectMapper;
-    }
 
     /**
      * web 请求的转换器，自动识别配置的spring bean
